i386/cpuid: Populate KVM_CPUID_VENDOR_FEATURE CPUID leaf
PVM, being a PV variant of KVM, utilizes a new CPUID leaf 0x40000002
within the KVM class, known as KVM_CPUID_VENDOR_FEATURE, to expose PVM
support for guests. However, QEMU does not utilize the CPUID information
obtained from KVM and always sets the maximum supported leaf of the KVM
class as 0x40000001. Therefore, in order to enable PVM to run with QEMU,
it is necessary to detect the new CPUID leaf and populate this CPUID
information during vCPU initialization.

Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>